+++ ios/Classes/FlutterEnvPlugin.swift
@@ -2,18 +2,5 @@ import Flutter
 import UIKit
 
 public class FlutterEnvPlugin: NSObject, FlutterPlugin {
-  public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "flutter_env", binaryMessenger: registrar.messenger())
-    let instance = FlutterEnvPlugin()
-    registrar.addMethodCallDelegate(instance, channel: channel)
-  }
-
-  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-    switch call.method {
-    case "getPlatformVersion":
-      result("iOS " + UIDevice.current.systemVersion)
-    default:
-      result(FlutterMethodNotImplemented)
-    }
-  }
+  public static func register(with registrar: FlutterPluginRegistrar) {}
 }